Add files via upload
Corresponds to an article introducing parents and kids to using ciphers and signaling. The activities utilize word search puzzles generated from online sites and the use of PowerShell. Parents can work with their child on running the script and how plaintext turns into ciphertext and back. The method may vary and the parent may choose to augment the lesson with an encrypt/decrypt function as outlined in the script comment as an example. The most basic lesson is the use of 1:1 cipher encodings similar to a Caesar cipher where the cipher is the method itself of the finding the letter position and back.